+++ Data/Retail/ExportedCalendarEvents.lua
@@ -24,7 +24,7 @@ local a = addon; -- Saves some characters and file size as we use this a lot in 
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2024-04-27 15-12-31 ]] --
+-- [[ Exported at 2024-05-09 09-40-18 ]] --
 tasks = {
     {N, 141, 917, 236705, a.L["Feast of Winter Veil"]}, -- Feast of Winter Veil
     {N, 181, 911, 237000, a.L["Noblegarden"]}, -- Noblegarden
@@ -56,12 +56,12 @@ tasks = {
     {N, 1430, 1560, 1100022, a.L["Pandaria Cup"]}, -- Pandaria Cup
     {N, 1431, 1561, 1100022, a.L["Broken Isles Cup"]}, -- Broken Isles Cup
     {N, 1462, 1563, 1061040, a.L["Hearthstone's 10th Anniversary"]}, -- Hearthstone's 10th Anniversary
-    {N, 1495, 1578, 5754532, a.L["Pandaria: Remix"], {1502, 1503, 1507}}, -- Pandaria: Remix
-    {N, 1502, 1578, 5754532, a.L["Pandaria: Remix"], {1495, 1503, 1507}}, -- Pandaria: Remix
-    {N, 1503, 1578, 5754532, a.L["Pandaria: Remix"], {1495, 1502, 1507}}, -- Pandaria: Remix
-    {N, 1507, 1578, 5754532, a.L["Pandaria: Remix"], {1495, 1502, 1503}}, -- Pandaria: Remix
     {N, 1500, 1173, 133783, a.L["WoW's 20th Anniversary"]}, -- WoW's 20th Anniversary
     {N, 1501, 1173, 133783, a.L["WoW's 21st Anniversary"]}, -- WoW's 21st Anniversary
+    {N, 1514, 1578, 5754532, a.L["WoW Remix: Mists of Pandaria"], {1515, 1516, 1525}}, -- WoW Remix: Mists of Pandaria
+    {N, 1515, 1578, 5754532, a.L["WoW Remix: Mists of Pandaria"], {1514, 1516, 1525}}, -- WoW Remix: Mists of Pandaria
+    {N, 1516, 1578, 5754532, a.L["WoW Remix: Mists of Pandaria"], {1514, 1515, 1525}}, -- WoW Remix: Mists of Pandaria
+    {N, 1525, 1578, 5754532, a.L["WoW Remix: Mists of Pandaria"], {1514, 1515, 1516}}, -- WoW Remix: Mists of Pandaria
 };
 
 function exportedCalendarEvents.InjectDynamicOptions()
@@ -90,8 +90,8 @@ function exportedCalendarEvents.InjectDynamicOptions()
     KrowiAF_RegisterEventOptions("Calendar", {1430}, a.L["Pandaria Cup"], a.L["Other"], 2);
     KrowiAF_RegisterEventOptions("Calendar", {1431}, a.L["Broken Isles Cup"], a.L["Other"], 2);
     KrowiAF_RegisterEventOptions("Calendar", {1462}, a.L["Hearthstone's 10th Anniversary"], a.L["Other"], 2);
-    KrowiAF_RegisterEventOptions("Calendar", {1495, 1502, 1503, 1507}, a.L["Pandaria: Remix"], a.L["Other"], 2);
     KrowiAF_RegisterEventOptions("Calendar", {1500}, a.L["WoW's 20th Anniversary"], a.L["Other"], 2);
     KrowiAF_RegisterEventOptions("Calendar", {1501}, a.L["WoW's 21st Anniversary"], a.L["Other"], 2);
-    KrowiAF_RegisterDeSelectAllEventOptions("Calendar", {479, 374, 514, 1262, 1395, 1396, 1398, 1399, 1397, 1400, 1407, 1425, 1458, 1459, 1460, 1429, 1430, 1431, 1462, 1495, 1502, 1503, 1507, 1500, 1501}, a.L["Other"]);
+    KrowiAF_RegisterEventOptions("Calendar", {1514, 1515, 1516, 1525}, a.L["WoW Remix: Mists of Pandaria"], a.L["Other"], 2);
+    KrowiAF_RegisterDeSelectAllEventOptions("Calendar", {479, 374, 514, 1262, 1395, 1396, 1398, 1399, 1397, 1400, 1407, 1425, 1458, 1459, 1460, 1429, 1430, 1431, 1462, 1500, 1501, 1514, 1515, 1516, 1525}, a.L["Other"]);
 end